backend/requirements.txt
##########################################################
#  Backend runtime deps (arm64 / x86, CPU-only)          #
##########################################################

# ─── ASGI stack ─────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-multipart>=0.0.17
aiofiles>=24.1.0

# ─── Media download / editing ────────────────────────────
yt-dlp>=2024.12.6
moviepy>=1.0.3
pydub>=0.25.1

# ─── ML models (CPU wheels) ────────────────────────────
# Install torch first: pip install torch --index-url https://download.pytorch.org/whl/cpu
torch>=2.5.0
openai-whisper>=20240930
demucs>=4.0.1

# ─── NLP / lyrics / fuzzy match ────────────────────────
rapidfuzz>=3.10.0
lyricsgenius>=3.0.1

# ─── Audio analysis ───────────────────────────────────
librosa>=0.10.1

# ─── HTTP / Misc ──────────────────────────────────────
httpx>=0.28.0
python-dotenv>=1.0.1
